refactor: reorganize Tripletex API structure by moving resources to endpoints and remove outdated files
