+++ tests/integration/tripletex_resources/endpoints/__init__.py
@@ -1,0 +1,11 @@
+from .companies import TripletexCompany
+from .countries import TripletexCountries
+from .ledger_group import LedgerGroup
+from .suppliers import TripletexSuppliers
+
+__all__: list[str] = [
+    "TripletexCompany",
+    "TripletexSuppliers",
+    "TripletexCountries",
+    "LedgerGroup",
+]
