+++ tests/integration/tripletex_resources/endpoints/ledger_group/__init__.py
@@ -1,0 +1,5 @@
+from .group import LedgerGroup
+
+__all__: list[str] = [
+    "LedgerGroup",
+]
