+++ tests/integration/tripletex_resources/endpoints/ledger_group/voucher_group/__init__.py
@@ -1,0 +1,5 @@
+from .group import VoucherGroup
+
+__all__: list[str] = [
+    "VoucherGroup",
+]
